+++ test/chat_web/live/main_live/index_test.exs
@@ -5,7 +5,7 @@ defmodule ChatWeb.MainLive.IndexTest do
   import Phoenix.LiveViewTest
 
   alias Chat.Admin.MediaSettings
-  alias Chat.{AdminDb, AdminRoom, Db, Dialogs, Rooms, RoomsBroker, User, UsersBroker}
+  alias Chat.{AdminDb, AdminRoom, Card, Db, Dialogs, Rooms, RoomsBroker, User, UsersBroker}
   alias Chat.Dialogs.PrivateMessage
   alias Chat.{Identity, Messages, RoomInviteIndex}
   alias Chat.Sync.{CargoRoom, UsbDriveDumpRoom}
@@ -40,14 +40,31 @@ defmodule ChatWeb.MainLive.IndexTest do
       another_tab: another_tab,
       conn: conn
     } do
-      %{view: current_tab_view} = current_tab
+      %{view: current_tab_view, socket: %{assigns: %{me: me}}} = current_tab
       %{view: another_tab_view} = another_tab
-      %{view: inviter_view} = prepare_view(%{conn: conn}) |> create_and_open_room("private")
 
-      inviter_view |> element("#roomInviteButton") |> IO.inspect() |> render_click()
-      # inviter_view |> element("#room-invite-list") |> IO.inspect
+      %{view: inviter_view, socket: %{assigns: %{room_identity: room_identity, me: inviter}}} =
+        prepare_view(%{conn: conn}) |> create_and_open_room("private")
 
-      IO.inspect(inviter_view)
+      my_hash = me |> Card.from_identity() |> then(& &1.hash)
+      inviter_view |> element("#roomInviteButton") |> render_click()
+      inviter_view |> element("#user-#{my_hash} a") |> render_click()
+
+      inviter_hash = inviter |> Card.from_identity() |> then(& &1.hash)
+      current_tab_view |> element("#dialog-list li#user-#{inviter_hash}") |> render_click()
+
+      current_tab_view |> element(".acceptInviteButton") |> render_click()
+
+      render_hook(current_tab_view, "room/sync-stored", %{
+        "key" => Identity.priv_key_to_string(room_identity),
+        "room_count" => 1
+      })
+
+      %{socket: %{assigns: current_view_state}} = reload_view(%{view: current_tab_view})
+      %{socket: %{assigns: another_view_state}} = reload_view(%{view: another_tab_view})
+
+      assert current_view_state.joined_rooms == another_view_state.joined_rooms
+      assert current_view_state.room_count_to_backup == another_view_state.room_count_to_backup
     end
   end